+++ feature/statistics-data/src/main/java/tmg/flashback/data/models/stats/Standings.kt
@@ -1,25 +1,25 @@
 package tmg.flashback.data.models.stats
 
-typealias ConstructorStandingsRound = Map<String, Triple<Constructor, Map<String, Pair<Driver, Int>>, Int>> // Constructor, driver id -> points per driver, constructor points
+typealias ConstructorStandingsRound = Map<String, Triple<Constructor, Map<String, Pair<Driver, Double>>, Double>> // Constructor, driver id -> points per driver, constructor points
 
-typealias DriverStandingsRound = Map<String, Pair<RoundDriver, Int>>
+typealias DriverStandingsRound = Map<String, Pair<RoundDriver, Double>>
 
 typealias ConstructorStandings = List<SeasonStanding<Constructor>>
 
 typealias DriverStandings = List<SeasonStanding<Driver>>
 
 data class RoundConstructorStandings(
-    val points: Int,
+    val points: Double,
     val constructor: Constructor
 )
 
 data class RoundDriverStandings(
-    val points: Int,
+    val points: Double,
     val driver: RoundDriver
 )
 
 data class SeasonStanding<T>(
     val item: T,
-    val points: Int,
+    val points: Double,
     val position: Int
 )
